Fix book reducer to add and remove
